Separate classes in single-class file #14
Game and Graphics WIP
