+++ src/lib/game/src/index.ts
@@ -1,0 +1,2 @@
+export * from './game';
+export * from './gameobject';
